new generic Component class that all other parts of sharpy should inherit. _client renamed to "client" in acts / plans in order to have a similar format managers were using.
