+++ sharpy/general/__init__.py
@@ -1,0 +1,1 @@
+# Left empty on purpose
