+++ sharpy/managers/manager_base.py
@@ -8,33 +8,14 @@ import sc2
 from sc2.client import Client
 from typing import TYPE_CHECKING
 
+from sharpy.general.component import Component
+
 if TYPE_CHECKING:
     from sharpy.knowledges import Knowledge, KnowledgeBot
     from sharpy.managers import UnitCacheManager, UnitValue
 
 
-class ManagerBase(ABC):
-    ai: "KnowledgeBot"
-    knowledge: "Knowledge"
-    unit_values: "UnitValue"
-    cache: "UnitCacheManager"
-    client: Client
-
-    def __init__(self):
-        self._debug: bool = False
-
-    @property
-    def debug(self):
-        return self._debug and self.knowledge.debug
-
-    async def start(self, knowledge: "Knowledge"):
-        self.knowledge = knowledge
-        self._debug = self.knowledge.config["debug"].getboolean(type(self).__name__)
-        self.ai = knowledge.ai
-        self.client = self.ai._client
-        self.cache = knowledge.unit_cache
-        self.unit_values = knowledge.unit_values
-
+class ManagerBase(ABC, Component):
     @abstractmethod
     async def update(self):
         pass
